swap canEdit guard with a resolver
